+++ src/main/java/domain/game/Name.java
@@ -1,11 +1,11 @@
-package domain;
+package domain.game;
 
 import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public record Name(String name) {
-    
+
     private static final Pattern pattern = Pattern.compile("^[^가-힣a-zA-Z-_0-9].*$");
     private static final int MAX_NAME_LENGTH = 5;
