+++ src/test/java/domain/card/CardNumberTest.java
@@ -1,12 +1,11 @@
-package domain;
+package domain.card;
 
-import domain.constant.CardNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class CardNumberTest {
-    
+
     @Test
     @DisplayName("카드 숫자에 맞는 점수를 반환한다.")
     void getScore() {
