+++ src/test/java/domain/card/CardTest.java
@@ -1,13 +1,11 @@
-package domain;
+package domain.card;
 
-import domain.constant.CardNumber;
-import domain.constant.CardType;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 public class CardTest {
-    
+
     @Test
     @DisplayName("생성자의 인자가 null일 경우 예외를 발생시킨다.")
     void createCard() {
