+++ src/test/java/domain/card/DeckTest.java
@@ -1,16 +1,14 @@
-package domain;
+package domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.constant.CardNumber;
-import domain.constant.CardType;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class DeckTest {
-    
+
     @Test
     @DisplayName("가장 마지막에 있는 카드를 반환한다.")
     void draw() {
