+++ src/test/java/domain/card/HandTest.java
@@ -1,7 +1,5 @@
-package domain;
+package domain.card;
 
-import domain.constant.CardNumber;
-import domain.constant.CardType;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
