+++ src/test/java/domain/game/DealerTest.java
@@ -1,14 +1,16 @@
-package domain;
+package domain.game;
 
-import domain.constant.CardNumber;
-import domain.constant.CardType;
+import domain.card.Card;
+import domain.card.CardNumber;
+import domain.card.CardType;
+import domain.card.Deck;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class DealerTest {
-    
+
     @Test
     @DisplayName("딜러는 자신의 패를 반환할 수 있다.")
     void getHand() {
@@ -23,7 +25,7 @@ class DealerTest {
         Dealer dealer = new Dealer();
         Card one = new Card(CardType.CLOVER, CardNumber.ACE);
         Card two = new Card(CardType.CLOVER, CardNumber.TWO);
-        Deck deck = Deck.withCustomCards(one, two);
+        Deck deck = TestDeck.withCustomCards(one, two);
         dealer.pickCards(deck, 2);
         List<Card> cards = dealer.getCards();
         Assertions.assertThat(cards)
@@ -35,7 +37,7 @@ class DealerTest {
     @DisplayName("딜러는 Hit 당 카드를 한장 뽑을 수 있다.")
     void takeCard() {
         Dealer dealer = new Dealer();
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.SPADE, CardNumber.TEN),
                 new Card(CardType.SPADE, CardNumber.FIVE),
                 new Card(CardType.SPADE, CardNumber.ACE),
@@ -48,7 +50,7 @@ class DealerTest {
     @DisplayName("딜러의 점수 합계를 반환한다.")
     void getTotalScore() {
         Dealer dealer = new Dealer();
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.SPADE, CardNumber.TEN),
                 new Card(CardType.SPADE, CardNumber.ACE));
         dealer.hit(deck);
@@ -61,7 +63,7 @@ class DealerTest {
     @DisplayName("Bust이면 true를 반환한다.")
     void isBustTrue() {
         Dealer dealer = new Dealer();
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.SPADE, CardNumber.TEN),
                 new Card(CardType.SPADE, CardNumber.SIX),
                 new Card(CardType.SPADE, CardNumber.TEN));
@@ -76,7 +78,7 @@ class DealerTest {
     @DisplayName("Bust가 아니면 false를 반환한다.")
     void isBustFalse() {
         Dealer dealer = new Dealer();
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.SPADE, CardNumber.TEN),
                 new Card(CardType.SPADE, CardNumber.TEN));
         dealer.hit(deck);
