+++ src/test/java/domain/game/NameTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -8,7 +8,7 @@ import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 public class NameTest {
-    
+
     @Test
     @DisplayName("생성자로 받은 이름 문자열을 반환한다")
     void getName() {
