+++ src/test/java/domain/game/PlayerTest.java
@@ -1,7 +1,10 @@
-package domain;
+package domain.game;
 
-import domain.constant.CardNumber;
-import domain.constant.CardType;
+import domain.card.Card;
+import domain.card.CardNumber;
+import domain.card.CardType;
+import domain.card.Deck;
+import domain.money.Money;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -23,7 +26,7 @@ class PlayerTest {
         Player player = new Player(new Name("test"));
         Card one = new Card(CardType.CLOVER, CardNumber.ACE);
         Card two = new Card(CardType.CLOVER, CardNumber.TWO);
-        Deck deck = Deck.withCustomCards(one, two);
+        Deck deck = TestDeck.withCustomCards(one, two);
         player.pickCards(deck, 2);
         List<Card> cards = player.getCards();
         Assertions.assertThat(cards)
@@ -36,7 +39,7 @@ class PlayerTest {
     void takeCard() {
         Player player = new Player(new Name("test"));
         Card card = new Card(CardType.SPADE, CardNumber.TEN);
-        Deck deck = Deck.withCustomCards(card);
+        Deck deck = TestDeck.withCustomCards(card);
         player.hit(deck);
         List<Card> cards = player.getCards();
         Assertions.assertThat(cards).isEqualTo(List.of(card));
@@ -53,7 +56,7 @@ class PlayerTest {
     @DisplayName("게이머의 점수 합계를 반환한다.")
     void getTotalScore() {
         Player player = new Player(new Name("test"));
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.SPADE, CardNumber.TEN));
         player.hit(deck);
         Assertions.assertThat(player.getTotalScore()).isEqualTo(10);
@@ -63,7 +66,7 @@ class PlayerTest {
     @DisplayName("Bust이면 true를 반환한다.")
     void isBustTrue() {
         Player player = new Player(new Name("test"));
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.SPADE, CardNumber.TEN),
                 new Card(CardType.SPADE, CardNumber.TEN),
                 new Card(CardType.SPADE, CardNumber.TEN));
@@ -79,7 +82,7 @@ class PlayerTest {
     @DisplayName("Bust가 아니면 false를 반환한다.")
     void isBustFalse() {
         Player player = new Player(new Name("test"));
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.SPADE, CardNumber.TEN),
                 new Card(CardType.SPADE, CardNumber.TEN));
         player.pickCards(deck, 2);
@@ -91,7 +94,7 @@ class PlayerTest {
     void betWinNormalCase() {
         Player player = new Player(new Name("test"), new Money(100));
         Dealer dealer = new Dealer();
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.SPADE, CardNumber.THREE),
                 new Card(CardType.SPADE, CardNumber.TWO),
                 new Card(CardType.DIAMOND, CardNumber.TEN),
@@ -106,7 +109,7 @@ class PlayerTest {
     void betLoseNormalCase() {
         Player player = new Player(new Name("test"), new Money(100));
         Dealer dealer = new Dealer();
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.DIAMOND, CardNumber.TEN),
                 new Card(CardType.DIAMOND, CardNumber.TEN),
                 new Card(CardType.SPADE, CardNumber.TWO),
@@ -122,7 +125,7 @@ class PlayerTest {
     void betWinBlackJackCase() {
         Player player = new Player(new Name("test"), new Money(100));
         Dealer dealer = new Dealer();
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.DIAMOND, CardNumber.TEN),
                 new Card(CardType.DIAMOND, CardNumber.TEN),
                 new Card(CardType.SPADE, CardNumber.ACE),
@@ -138,7 +141,7 @@ class PlayerTest {
     void betLoseBlackJackCase() {
         Player player = new Player(new Name("test"), new Money(100));
         Dealer dealer = new Dealer();
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.DIAMOND, CardNumber.ACE),
                 new Card(CardType.DIAMOND, CardNumber.KING),
                 new Card(CardType.SPADE, CardNumber.KING),
@@ -154,7 +157,7 @@ class PlayerTest {
     void betBothBlackJackCase() {
         Player player = new Player(new Name("test"), new Money(100));
         Dealer dealer = new Dealer();
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.DIAMOND, CardNumber.ACE),
                 new Card(CardType.DIAMOND, CardNumber.KING),
                 new Card(CardType.SPADE, CardNumber.ACE),
@@ -170,7 +173,7 @@ class PlayerTest {
     void betSameScoreCase() {
         Player player = new Player(new Name("test"), new Money(100));
         Dealer dealer = new Dealer();
-        Deck deck = Deck.withCustomCards(
+        Deck deck = TestDeck.withCustomCards(
                 new Card(CardType.DIAMOND, CardNumber.KING),
                 new Card(CardType.DIAMOND, CardNumber.JACK),
                 new Card(CardType.SPADE, CardNumber.KING),
